apply validation on creating a privilage

+++ access/authorization_class/privilage.py
@@ -1,4 +1,6 @@
 import sys
+
+import PyQt5
 from PyQt5.QtCore import pyqtSlot
 from PyQt5.QtWidgets import QApplication, QDialog, QTableWidgetItem
 from PyQt5.uic import loadUi
@@ -10,6 +12,9 @@ import mysql.connector
 import os
 import sys
 from pathlib import Path
+
+from data_connection.h1pos import db1
+
 
 class CL_privilage(QtWidgets.QDialog):
     dirname=''
@@ -45,7 +50,6 @@ class CL_privilage(QtWidgets.QDialog):
         rows = []
         for idx in self.w1.selectionModel().selectedRows():
             rows.append( idx.row() )
-
         #print(rows)
         for row in rows:
             self.w1.removeRow( row )
@@ -58,34 +62,78 @@ class CL_privilage(QtWidgets.QDialog):
         self.roleName = self.CMB_roleName.currentText()
         self.formName = self.CMB_formName.currentText()
         self.actionName = self.CMB_actionName.currentText()
+        self.actionId = self.LB_actionId.text()
         self.formItemName = self.CMB_formItemName.currentText()
-        #print (self.formItemName )
-        rowPosition = self.w1.rowCount()
-        #rowPosition = rowPosition+1
-        #print(rowPosition)
-        self.w1.setRowCount( rowPosition )
-        self.w1.insertRow( self.w1.rowCount())
 
-        self.w1.setItem( rowPosition, 0, QTableWidgetItem( str(self.roleName) ) )
-        self.w1.setItem( rowPosition, 1, QTableWidgetItem( str( self.role ) ) )
-        self.w1.setItem( rowPosition, 2, QTableWidgetItem( str( self.formName ) ) )
-        self.w1.setItem( rowPosition, 3, QTableWidgetItem( str( self.form ) ) )
-        self.w1.setItem( rowPosition, 4, QTableWidgetItem( str( self.actionName ) ) )
-        self.w1.setItem( rowPosition, 5, QTableWidgetItem( str( self.formItemName ) ) )
+        x = self.FN_CHECK_DB_AVAILABILITY( self.role, self.form, self.actionId, self.formItem )
 
-    def FN_LOAD_MODFIY(self):
-        filename = self.dirname + '/modifyPrivilage.ui'
-        loadUi( filename, self )
-        #loadUi('../Presentation/modifyPrivilage.ui', self)
+        if x == True:
+            QtWidgets.QMessageBox.warning( self, "Error", "Privilage already exists" )
+        else:
+            print("in else")
+            if self.FN_CHECK_TABLE_WIDGET_AVAILABILITY(self.role, self.form, self.actionId, self.formItem) == True :
+                QtWidgets.QMessageBox.warning( self, "Error", "You already entered this Priviliage in the grid" )
+            else:
 
-        self.LE_id.textChanged.connect(self.FN_GET_PRIV)
-        self.BTN_modifyPrivilage.clicked.connect(self.FN_MODIFY_PRIV)
-        self.CMB_roleName.currentIndexChanged.connect(self.FN_GET_ROLEID)
-        self.CMB_formName.currentIndexChanged.connect(self.FN_GET_FORMID)
-        self.CMB_actionName.currentIndexChanged.connect(self.FN_GET_ACTIONID)
-        # self.FN_GET_ROLEID()
-        # self.FN_GET_FORMID()
-        # self.FN_GET_ACTIONID()
+                rowPosition = self.w1.rowCount()
+                self.w1.setRowCount( rowPosition )
+                self.w1.insertRow( self.w1.rowCount())
+
+                self.w1.setItem( rowPosition, 0, QTableWidgetItem( str(self.roleName) ) )
+                self.w1.setItem( rowPosition, 1, QTableWidgetItem( str( self.role ) ) )
+                self.w1.setItem( rowPosition, 2, QTableWidgetItem( str( self.formName ) ) )
+                self.w1.setItem( rowPosition, 3, QTableWidgetItem( str( self.form ) ) )
+                self.w1.setItem( rowPosition, 4, QTableWidgetItem( str( self.actionName ) ) )
+                self.w1.setItem( rowPosition, 5, QTableWidgetItem( str( self.formItemName ) ) )
+    def FN_CHECK_TABLE_WIDGET_AVAILABILITY(self,var11, var2, var3, var4):
+
+        mycursor = db1.connect( self )
+        allRows = self.w1.rowCount()
+
+        for row in range( 0, allRows ):
+
+            sql_select_query = "SELECT ACTION_ID FROM SYS_PRINT_EXPORT_LOOKUP WHERE ACTION_DESC = %s"
+            x = (self.w1.item( row, 4 ).text(),)
+            mycursor.execute( sql_select_query, x )
+
+            myresult = mycursor.fetchone()
+            if mycursor.rowcount > 0:
+                actionId = myresult[0]
+
+            formItemName = self.w1.item( row, 5 )
+            sql_select_query = "SELECT ITEM_ID FROM SYS_FORM_ITEM WHERE ITEM_DESC = %s"
+            x = (formItemName.text(),)
+            mycursor.execute( sql_select_query, x )
+
+            myresult = mycursor.fetchone()
+            if mycursor.rowcount > 0:
+                formItemId = myresult[0]
+
+            if var11 == self.w1.item( row, 1 ).text() and var2 ==self.w1.item( row, 3 ).text()and var3 == actionId and var4 == formItemId:
+               return True
+
+        # if chk:
+        #     return True
+
+
+
+    def FN_CHECK_DB_AVAILABILITY(self,var11, var2, var3, var4):
+        #print(var11, var2, var3, var4)
+        mycursor = db1.connect(self)
+        sqlStat= "Select  *      " \
+                 "from SYS_PRIVILEGE p inner join SYS_PRIVILEG_ITEM pi on p.PRIV_ID= pi.PRIV_ID  " \
+                 "and p.FORM_ID = pi.FORM_ID                 " \
+                 "where  p.ROLE_ID = "+var11+" and p.FORM_ID = "+var2+" and p.ACTION_ID = "+var3+" and pi.ITEM_ID="+var4+""
+        #print(sqlStat)
+        mycursor.execute( sqlStat )
+
+        records = mycursor.fetchall()
+
+        if mycursor.rowcount > 0:
+            return True
+        else:
+            return False
+
     def FN_GET_ACTIONS(self):
         connection = mysql.connector.connect(host='localhost', database='PosDB'
                                              , user='root', password='password', port='3306')
@@ -218,59 +266,68 @@ class CL_privilage(QtWidgets.QDialog):
 
         connection.close()
         mycursor.close()
-    def FN_GET_PRIV(self):
+    # def FN_GET_PRIV(self):
+    #
+    #     self.id = self.LE_id.text()
+    #     self.FN_GET_ROLES()
+    #     self.FN_GET_FORMS()
+    #     self.FN_GET_ACTIONS()
+    #     # self.FN_GET_ROLEID()
+    #     # self.FN_GET_FORMID()
+    #     # self.FN_GET_ACTIONID()
+    #     connection = mysql.connector.connect(host='localhost', database='PosDB'
+    #                                          , user='root', password='password', port='3306')
+    #     mycursor = connection.cursor()
+    #     sql_select_query = "select * from SYS_PRIVILEGE where PRIV_ID = %s"
+    #     x = (self.id,)
+    #     mycursor.execute(sql_select_query, x)
+    #     record = mycursor.fetchone()
+    #
+    #     self.LB_roleId.setText(record[1])
+    #     self.LB_formId.setText(record[2])
+    #     self.LB_actionId.setText(record[3])
+    #    # self.FN_GET_ROLENAME()
+    #     #self.CMB_roleName.setCurrentIndex(self,1)
+    #     connection.close()
+    #     mycursor.close()
+    #
+    #     print(mycursor.rowcount, "record retrieved.")
 
-        self.id = self.LE_id.text()
-        self.FN_GET_ROLES()
-        self.FN_GET_FORMS()
-        self.FN_GET_ACTIONS()
-        # self.FN_GET_ROLEID()
-        # self.FN_GET_FORMID()
-        # self.FN_GET_ACTIONID()
-        connection = mysql.connector.connect(host='localhost', database='PosDB'
-                                             , user='root', password='password', port='3306')
-        mycursor = connection.cursor()
-        sql_select_query = "select * from SYS_PRIVILEGE where PRIV_ID = %s"
-        x = (self.id,)
-        mycursor.execute(sql_select_query, x)
-        record = mycursor.fetchone()
-
-        self.LB_roleId.setText(record[1])
-        self.LB_formId.setText(record[2])
-        self.LB_actionId.setText(record[3])
-       # self.FN_GET_ROLENAME()
-        #self.CMB_roleName.setCurrentIndex(self,1)
-        connection.close()
-        mycursor.close()
-
-        print(mycursor.rowcount, "record retrieved.")
-    def FN_MODIFY_PRIV(self):
-        self.id = self.LE_id.text()
-        self.role = self.LB_roleId.text()
-        self.form = self.LB_formId.text()
-        self.action = self.LB_actionId.text()
-
-        #         connection = mysql.connector.connect(host='localhost',database='test',user='shelal',password='2ybQvkZbNijIyq2J',port='3306')
-        connection = mysql.connector.connect(host='localhost', database='PosDB'
-                                             , user='root', password='password', port='3306')
-
-        mycursor = connection.cursor()
-
-
-        sql = "UPDATE SYS_PRIVILEGE  set ROLE_ID= %s ,  FORM_ID= %s  ,  ACTION_ID= %s  where PRIV_ID= %s "
-
-        val = (self.role , self.form,   self.action, self.id)
-
-        mycursor.execute(sql, val)
-        # mycursor.execute(sql)
-        connection.commit()
-
-        connection.close()
-        mycursor.close()
-
-        print(mycursor.rowcount, "record Modified.")
-
-        self.close()
+    # def FN_LOAD_MODFIY(self):
+    #     filename = self.dirname + '/modifyPrivilage.ui'
+    #     loadUi( filename, self )
+    #     self.LE_id.textChanged.connect( self.FN_GET_PRIV )
+    #     self.BTN_modifyPrivilage.clicked.connect( self.FN_MODIFY_PRIV )
+    #     self.CMB_roleName.currentIndexChanged.connect( self.FN_GET_ROLEID )
+    #     self.CMB_formName.currentIndexChanged.connect( self.FN_GET_FORMID )
+    #     self.CMB_actionName.currentIndexChanged.connect( self.FN_GET_ACTIONID )
+    # def FN_MODIFY_PRIV(self):
+    #     self.id = self.LE_id.text()
+    #     self.role = self.LB_roleId.text()
+    #     self.form = self.LB_formId.text()
+    #     self.action = self.LB_actionId.text()
+    #
+    #     #         connection = mysql.connector.connect(host='localhost',database='test',user='shelal',password='2ybQvkZbNijIyq2J',port='3306')
+    #     connection = mysql.connector.connect(host='localhost', database='PosDB'
+    #                                          , user='root', password='password', port='3306')
+    #
+    #     mycursor = connection.cursor()
+    #
+    #
+    #     sql = "UPDATE SYS_PRIVILEGE  set ROLE_ID= %s ,  FORM_ID= %s  ,  ACTION_ID= %s  where PRIV_ID= %s "
+    #
+    #     val = (self.role , self.form,   self.action, self.id)
+    #
+    #     mycursor.execute(sql, val)
+    #     # mycursor.execute(sql)
+    #     connection.commit()
+    #
+    #     connection.close()
+    #     mycursor.close()
+    #
+    #     print(mycursor.rowcount, "record Modified.")
+    #
+    #     self.close()
     def FN_DISPLAY_PRIVILAGE(self):
         self.w1.clear()
         self.w1.setRowCount(0)
@@ -303,6 +360,9 @@ class CL_privilage(QtWidgets.QDialog):
         # self.w1.setItem(0, 0, QTableWidgetItem("Name"))
         connection.close()
 
+        self.w1.setEditTriggers(QtWidgets.QTableWidget.NoEditTriggers)
+        header_labels = ['Role Name', 'Role Id', 'Form Name', 'Form Id', 'Action Name','Form Item']
+        self.w1.setHorizontalHeaderLabels( header_labels )
 
     def FN_CREATE_PRIVILAGE(self):
         connection = mysql.connector.connect( host='localhost', database='PosDB'
